# app/Http/Requests/StoreAuctionRequest.php
<?php

namespace App\Http\Requests;

use Illuminate\Foundation\Http\FormRequest;

/**
 * @OA\Schema(
 *     title="StoreAuctionRequest",
 *     description="Store Auction Request",
 *     @OA\Xml(
 *         name="StoreAuctionRequest"
 *     )
 * )
 */
class StoreAuctionRequest extends FormRequest
{

    /**
     * Determine if the user is authorized to make this request.
     *
     * @return bool
     */
    public function authorize()
    {
        return true;
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array<string, mixed>
     */
    public function rules()
    {
        return [
            'book_id' => 'required',
            'price' => 'required',
            'enabled' => 'required',
            'quantity' => 'required',
        ];
    }
}